Crate gilrs_core

Modules

native_ev_codes
Provides the most common mappings of physical location of gamepad elements to their EvCodes. Some (or most) gamepads may use different mappings.

utils

Structs

AxisInfo
Holds information about expected axis range and deadzone.

EvCode
Platform specific representation of axis or button.

Event
Holds information about gamepad event.

FfDevice
Allow control of gamepad's force feedback.

Gamepad
Provides information about gamepad.

Gilrs
Struct used to manage gamepads and retrieve events.

Enums

Error
Error type which can be returned when creating Gilrs.

EventType
Gamepad event.

PowerInfo
State of device's power supply.

Constants

IS_Y_AXIS_REVERSED
True, if Y axis of sticks commonly points downwards.