gilrs_core
Struct Gamepad

pub struct Gamepad { /* private fields */ }

Provides information about gamepad.

Implementations

impl Gamepad

pub fn name(&self) -> &str

Returns name of gamepad.

pub fn is_connected(&self) -> bool

Returns true if gamepad is connected.

pub fn uuid(&self) -> [u8; 16]

Returns UUID that represents gamepad model.

Returned UUID should be the same as SLD2 uses. If platform does not provide any method to distinguish between gamepad models, nil UUID is returned.

It is recommended to process with the UUID crate. Use Uuid::from_bytes method to create a Uuid from the returned bytes.

pub fn vendor_id(&self) -> Option<u16>

Returns the vendor ID, as assigned by the USB-IF, when available.

pub fn product_id(&self) -> Option<u16>

Returns the product ID, as assigned by the vendor, when available.

pub fn power_info(&self) -> PowerInfo

Returns device’s power supply state.

pub fn is_ff_supported(&self) -> bool

Returns true if force feedback is supported by device,

pub fn ff_device(&self) -> Option<FfDevice>

Creates FfDevice corresponding to this gamepad.

pub fn buttons(&self) -> &[EvCode]

Returns slice with EvCodes that may appear in button related events.

pub fn axes(&self) -> &[EvCode]

Returns slice with EvCodes that may appear in axis related events.

pub fn axis_info(&self, nec: EvCode) -> Option<&AxisInfo>

Returns information about specific axis. None may be returned if device doesn’t have axis with provided EvCode.